Split mesh generation into base class
