+++ TestAssets/main_test.tscn
@@ -58,11 +58,12 @@ albedo_texture = ExtResource("12_0ug2n")
 roughness_texture = ExtResource("14_y08bg")
 normal_enabled = true
 normal_texture = ExtResource("13_lye7l")
+uv1_scale = Vector3(4, 4, 4)
 
 [sub_resource type="BoxShape3D" id="12"]
 size = Vector3(11.484, 6.85648, 7.49257)
 
-[sub_resource type="Resource" id="Resource_o070k"]
+[sub_resource type="Resource" id="Resource_ch278"]
 script = ExtResource("15_pphx5")
 MatShaderType = 1
 mat_albedo_color_first = Color(0.6, 0.7, 0.65, 1)
@@ -90,11 +91,11 @@ mat_edge_fade = 0.25
 LodLod0Distance = 50.0
 _selectedShader = 1
 
-[sub_resource type="Curve3D" id="Curve3D_b6rac"]
+[sub_resource type="Curve3D" id="Curve3D_3foam"]
 resource_local_to_scene = true
 bake_interval = 0.05
 _data = {
-"points": PackedVector3Array(-3.54596, 0.142921, -4.43361, 3.54596, -0.142921, 4.43361, -16.2033, 2.55085, -39.3207, -1.92814, -0.301395, -3.25125, 1.92814, 0.301395, 3.25125, -7.00658, -0.415245, -24.7892, -0.758025, -0.153267, -2.91529, 0.758025, 0.153267, 2.91529, -2.14075, -0.345051, -7.42294, 3.28793, -0.0342751, -9.96773, -3.28793, 0.0342751, 9.96773, -2.02899, -0.923592, 9.13994, 2.35744, 0.0164261, -2.79135, -2.35744, -0.0164261, 2.79135, -7.86718, -0.817085, 23.175, 5.3857, -0.871243, -2.40109, -5.3857, 0.871243, 2.40109, -25.6885, 0.992832, 35.2528, 6.89437, -0.848625, -2.45419, -6.89437, 0.848625, 2.45419, -57.1386, 3.49468, 44.3398, 2.41856, -0.391121, -3.886, -2.41856, 0.391121, 3.886, -80.088, 8.47322, 57.3421, 1.22588, 0.369118, -3.23756, -1.22588, -0.369118, 3.23756, -85.9448, 9.15034, 72.5185),
+"points": PackedVector3Array(-3.54596, 0.142921, -4.43361, 3.54596, -0.142921, 4.43361, -16.2033, 2.55085, -39.3207, -1.92814, -0.301395, -3.25125, 1.92814, 0.301395, 3.25125, -7.00658, -0.415245, -24.7892, -0.758025, -0.153267, -2.91529, 0.758025, 0.153267, 2.91529, -2.14075, -0.345051, -7.42294, 3.28793, -0.0342751, -9.96773, -3.28793, 0.0342751, 9.96773, -2.03336, -0.810101, 9.07495, 2.35744, 0.0164261, -2.79135, -2.35744, -0.0164261, 2.79135, -7.86718, -0.817085, 23.175, 5.3857, -0.871243, -2.40109, -5.3857, 0.871243, 2.40109, -25.6885, 0.992832, 35.2528, 6.89437, -0.848625, -2.45419, -6.89437, 0.848625, 2.45419, -57.1386, 3.49468, 44.3398, 2.41856, -0.391121, -3.886, -2.41856, 0.391121, 3.886, -80.088, 8.47322, 57.3421, 1.22588, 0.369118, -3.23756, -1.22588, -0.369118, 3.23756, -85.9448, 9.15034, 72.5185),
 "tilts": PackedFloat32Array(0, 0, 0, 0, 0, 0, 0, 0, 0)
 }
 point_count = 9
@@ -140,7 +141,7 @@ transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, 0, -1.34785, 0)
 transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1.34785, 0)
 
 [node name="rock2" type="MeshInstance3D" parent="Boulders/Hill"]
-transform = Transform3D(-7.10392, -18.6411, 1.43002, -11.1647, 5.45689, 15.6708, -14.9962, 4.76791, -12.3443, 50.5136, 41.3095, -17.6291)
+transform = Transform3D(-2.88559, -19.386, -3.98221, -12.9147, -1.20486, 15.2237, -14.9962, 4.76791, -12.3443, 51.364, 41.3095, -18.4528)
 mesh = ExtResource("11_34ive")
 skeleton = NodePath("../..")
 surface_material_override/0 = SubResource("StandardMaterial3D_sn1j4")
@@ -164,7 +165,7 @@ skeleton = NodePath("../..")
 surface_material_override/0 = SubResource("StandardMaterial3D_sn1j4")
 
 [node name="rock7" type="MeshInstance3D" parent="Boulders/Hill"]
-transform = Transform3D(-16.2859, -10.9862, 3.75151, -1.20121, 8.02313, 18.2808, -11.5467, 14.6606, -7.19302, 38.0219, 38.5004, -34.1636)
+transform = Transform3D(-14.5932, -13.2173, -3.51317, -7.32887, 3.22018, 18.3281, -11.5467, 14.6606, -7.19302, 37.7799, 38.5004, -35.8373)
 mesh = ExtResource("11_34ive")
 skeleton = NodePath("../..")
 surface_material_override/0 = SubResource("StandardMaterial3D_sn1j4")
@@ -176,13 +177,13 @@ skeleton = NodePath("../..")
 surface_material_override/0 = SubResource("StandardMaterial3D_sn1j4")
 
 [node name="rock5" type="MeshInstance3D" parent="Boulders/Hill"]
-transform = Transform3D(-12.9947, -13.5999, 6.79557, -10.1283, 14.4111, 9.47299, -11.3382, 2.71357, -16.2506, 53.8979, 56.0384, -30.8889)
+transform = Transform3D(-12.9947, -13.5999, 6.79557, -10.1283, 14.4111, 9.47299, -11.3382, 2.71357, -16.2506, 50.0374, 56.0384, -29.7487)
 mesh = ExtResource("11_34ive")
 skeleton = NodePath("../..")
 surface_material_override/0 = SubResource("StandardMaterial3D_sn1j4")
 
 [node name="rock6" type="MeshInstance3D" parent="Boulders/Hill"]
-transform = Transform3D(-11.2419, -12.303, -11.0569, -15.2927, 12.8259, 1.27721, 6.30508, 9.17242, -16.6166, 55.4883, 61.0773, -35.1688)
+transform = Transform3D(-11.2419, -12.303, -11.0569, -12.11, 15.2281, -4.6316, 11.268, 4.0916, -16.0091, 45.1855, 59.8577, -33.822)
 mesh = ExtResource("11_34ive")
 skeleton = NodePath("../..")
 surface_material_override/0 = SubResource("StandardMaterial3D_sn1j4")
@@ -191,7 +192,7 @@ surface_material_override/0 = SubResource("StandardMaterial3D_sn1j4")
 transform = Transform3D(1.93763, 0, 0, 0, 1.93763, 0, 0, 0, 1.93763, -9.51027, 41.6634, 11.3927)
 
 [node name="Camera" type="Camera3D" parent="."]
-transform = Transform3D(0.677773, -0.256857, 0.688947, 0.100967, 0.960635, 0.258819, -0.728305, -0.10586, 0.677027, 44, 15, 30)
+transform = Transform3D(0.866025, 0.12941, -0.482963, 0, 0.965926, 0.258819, 0.5, -0.224144, 0.836516, -21.4558, 56.7776, 18.5822)
 
 [node name="FloatSpawner" type="Node3D" parent="Camera"]
 transform = Transform3D(1, 0, -2.98023e-08, 0, 1, 0, 2.98023e-08, 0, 1, 0, -1, 0)
@@ -209,6 +210,6 @@ shape = SubResource("12")
 [node name="RiverManager" type="Node3D" parent="."]
 transform = Transform3D(-0.0584003, 0, -0.998293, 0, 1, 0, 0.998293, 0, -0.0584003, 0.452626, 33.1977, -7.67319)
 script = ExtResource("13_dg81e")
-ShaderSettings = SubResource("Resource_o070k")
+ShaderSettings = SubResource("Resource_ch278")
 PointWidths = [6.33827, 7.97592, 7.08766, 5.21408, 7.80158, 6.03617, 6.81221, 7.42093, 6.06767]
-Curve = SubResource("Curve3D_b6rac")
+Curve = SubResource("Curve3D_3foam")
